Use Azure for macOS builds
We switched to CircleCI in this feedstock back when we were running into
CI timelimits on Travis CI. So CircleCI was a preferable option. However
now that most builds are on Azure and they provide ample time, it makes
sense to switch. Plus it is the conda-forge default at this point.